LPS-74449 Test CLI parsing as well
