+++ modules/util/css-builder/src/test/java/com/liferay/css/builder/ant/BuildCSSTaskTest.java
@@ -14,7 +14,7 @@
 
 package com.liferay.css.builder.ant;
 
-import com.liferay.css.builder.CSSBuilderTest;
+import com.liferay.css.builder.BaseCSSBuilderTestCase;
 import com.liferay.portal.kernel.util.StringUtil;
 
 import java.io.File;
@@ -33,7 +33,7 @@ import org.junit.Rule;
 /**
  * @author Andrea Di Giorgi
  */
-public class BuildCSSTaskTest extends CSSBuilderTest {
+public class BuildCSSTaskTest extends BaseCSSBuilderTestCase {
 
 	@Before
 	@Override
